Switch to row-based visual callbacks
This changes the behaviour of the callbacks to match how selections work
in Google Chart Tools and C3, i.e. by row, across all series (if
specific datum points are requested for selection, then the selection
only applies to the last one selected, rather than all in the row), and
to simplify HTML tables such that the entire row gets highlighted.